+++ src/main/kotlin/com/github/themartdev/intellijgleam/ide/ui/GleamSettingsConfigurable.kt
@@ -3,10 +3,12 @@ package com.github.themartdev.intellijgleam.ide.ui
 import com.github.themartdev.intellijgleam.GleamBundle
 import com.github.themartdev.intellijgleam.ide.common.ErlangSdkFinder
 import com.github.themartdev.intellijgleam.ide.common.GleamExecutableFinder
+import com.github.themartdev.intellijgleam.ide.common.JsRuntimeFinder
 import com.github.themartdev.intellijgleam.ide.lsp.GleamLspMode
 import com.github.themartdev.intellijgleam.ide.lsp.GleamServiceSettings
 import com.github.themartdev.intellijgleam.ide.ui.components.ErlangPathComboBox
 import com.github.themartdev.intellijgleam.ide.ui.components.GleamPathComboBox
+import com.github.themartdev.intellijgleam.ide.ui.components.JsRuntimePathComboBox
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ModalityState
 import com.intellij.openapi.options.BoundConfigurable
@@ -24,10 +26,12 @@ class GleamSettingsConfigurable(private val project: Project) :
 
     private val gleamPathComboBox = GleamPathComboBox(project)
     private val erlangPathComboBox = ErlangPathComboBox(project)
+    private val jsRuntimePathComboBox = JsRuntimePathComboBox(project)
 
     override fun createPanel() = panel {
         loadDetectedGleamPaths()
         loadDetectedErlangPaths()
+        loadDetectedJsRuntimePaths()
         group(GleamBundle.message("gleam.settings.configurable.group.sdk")) {
             lateinit var overrideCheckBox: Cell<JBCheckBox>
             row {
@@ -50,6 +54,15 @@ class GleamSettingsConfigurable(private val project: Project) :
                     settings::erlangPath.toMutableProperty()
                 )
             }.enabledIf(overrideCheckBox.selected)
+
+            row(GleamBundle.message("gleam.settings.configurable.sdk.jsRuntimePath")) {
+                cell(jsRuntimePathComboBox).align(AlignX.FILL).bind(
+                    { it.selectedPath ?: "" },
+                    { a, b -> a.selectedPath = b },
+                    settings::jsRuntimePath.toMutableProperty()
+                )
+            }.enabledIf(overrideCheckBox.selected)
+                .rowComment(GleamBundle.message("gleam.settings.configurable.sdk.jsRuntimePath.help"))
 
         }
         group(GleamBundle.message("gleam.settings.configurable.group.lsp")) {
@@ -83,6 +96,18 @@ class GleamSettingsConfigurable(private val project: Project) :
             ApplicationManager.getApplication().invokeLater({
                 detectedErlangPaths.forEach { path ->
                     erlangPathComboBox.addItem(path.path, path.version)
+                }
+            }, modalityState)
+        }
+    }
+
+    private fun loadDetectedJsRuntimePaths() {
+        val modalityState = ModalityState.current()
+        ApplicationManager.getApplication().executeOnPooledThread {
+            val detectedJsRuntimes = JsRuntimeFinder.findJsRuntimes()
+            ApplicationManager.getApplication().invokeLater({
+                detectedJsRuntimes.forEach { runtime ->
+                    jsRuntimePathComboBox.addItem(runtime.path, runtime.version)
                 }
             }, modalityState)
         }